+++ xelis_common/src/tokio/mod.rs
@@ -15,3 +15,47 @@ pub use tokio_with_wasm::*;
     target_os = "unknown"
 )))]
 pub use tokio::*;
+
+
+// Spawn a new task with a name
+// If the tokio_unstable feature is enabled, the task will be named
+#[inline(always)]
+#[cfg(not(all(
+    target_arch = "wasm32",
+    target_vendor = "unknown",
+    target_os = "unknown"
+)))]
+pub fn spawn_task<Fut, S: Into<String>>(name: S, future: Fut) -> task::JoinHandle<Fut::Output>
+where
+    Fut: std::future::Future + Send + 'static,
+    Fut::Output: Send + 'static,
+{
+    let name_str = name.into();
+    log::trace!("Spawning task: {}", name_str);
+    #[cfg(all(tokio_unstable, feature = "tracing"))]
+    {
+        let name = name_str.as_str();
+        task::Builder::new().name(name).spawn(future).expect(name)
+    }
+    #[cfg(not(all(tokio_unstable, feature = "tracing")))]
+    {
+        tokio::spawn(future)
+    }
+}
+
+// Spawn a new task with a name
+// Send trait is not required for wasm32
+#[cfg(all(
+    target_arch = "wasm32",
+    target_vendor = "unknown",
+    target_os = "unknown"
+))]
+pub fn spawn_task<Fut, S: Into<String>>(name: S, future: Fut) -> task::JoinHandle<Fut::Output>
+where
+    Fut: std::future::Future + 'static,
+    Fut::Output: 'static,
+{
+    let name_str = name.into();
+    log::trace!("Spawning wasm task: {}", name_str);
+    spawn(future)
+}
